requirements.txt
openai>=1.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pytest>=8.0.0
ruff>=0.8.0
mypy>=1.8.0
pytest-cov>=5.0.0
